Added Resource besides AIResource, made News and EducationalResource a Resource, not AIResource
