+++ src/config.toml
@@ -2,7 +2,7 @@
 
 # Information on which database to connect to
 [database]
-host = "172.17.0.2"
+host = "sqlserver"
 port = 3306
 database = "aiod"
 username = "root"
